+++ src/index.ts
@@ -1,14 +1,11 @@
 /**
  * NinePatchPlugin
  */
-export { NinePatchPlugin } from "./plugin/NinePatchPlugin";
-
-export { INinePatchCreator } from "./plugin/ninepatch/INinePatchCreator";
-
-export { INinePatchFactory } from "./plugin/ninepatch/INinePatchFactory";
-
 export { INinePatchConfig } from "./plugin/ninepatch/INinePatchConfig";
-
+export { INinePatchCreator } from "./plugin/ninepatch/INinePatchCreator";
+export { INinePatchFactory } from "./plugin/ninepatch/INinePatchFactory";
+export { INinePatchScene } from "./plugin/ninepatch/INinePatchScene";
 export { IPatchesConfig } from "./plugin/ninepatch/IPatchesConfig";
-
 export { NinePatch } from "./plugin/ninepatch/NinePatch";
+export { NinePatchCacheManager } from "./plugin/ninepatch/NinePatchCacheManager";
+export { NinePatchPlugin } from "./plugin/NinePatchPlugin";