+++ src/plugin/ninepatch/NinePatchCacheManager.ts
@@ -1,0 +1,3 @@
+export class NinePatchCacheManager extends Phaser.Cache.CacheManager {
+    public custom: { ninePatch: Phaser.Cache.BaseCache; [key: string]: Phaser.Cache.BaseCache };
+}